refactor: export 시 .js 확장자 제거
