+++ src/services/index.js
@@ -1,2 +1,2 @@
-export { default as Logger } from './logger.js'
-export { default as Storage } from './storage.js'
+export { default as Logger } from './logger'
+export { default as Storage } from './storage'
