+++ src/utils/index.js
@@ -1,3 +1,3 @@
-export { default as memoize } from './memoize.js'
-export { default as throttle } from './throttle.js'
-export { default as debounce } from './debounce.js'
+export { default as memoize } from './memoize'
+export { default as throttle } from './throttle'
+export { default as debounce } from './debounce'
